+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 #### Fixed
 - add missing _must_ keyword in expectation convention paragraph of README
 - add missing comment for QuickMock.support.backupPrefix
+- leading and trailing spaces are removed from expectation names during mock creation
 
 #### Added
 - shouldReceive expectation